+++ src/extensionDefaults.ts
@@ -3,7 +3,7 @@ export class ExtensionDefaults {
     public static readonly defaultEditorConfig = "coboleditor";
     public static readonly thisExtensionName = "bitlang.cobol";
 
-    public static readonly microFocusCOBOLExtension = "RocketSoftware.rocket-cobol";
-    public static readonly microFocusEnterpriseExtenstion = "RocketSoftware.rocket-enterprise";
+    public static readonly rocketCOBOLExtension = "RocketSoftware.rocket-cobol";
+    public static readonly rocketEnterpriseExtenstion = "RocketSoftware.rocket-enterprise";
     public static readonly microFocusCOBOLLanguageId = "cobol";
 }